+++ ci/release_notes.md
@@ -3,6 +3,7 @@
 
 # New Features
 - socat is directly executable due to a symlink
+- expose `h1-accept-payload-with-any-method` as `ha_proxy.always_allow_body_http10`
 
 # Upgrades
 - HAProxy 2.5.4 -> 2.5.7